MANTA-2124: make deploy-coal-manta.sh re-runnable

+++ devtools/deploy-coal-manta.sh
@@ -8,11 +8,13 @@ set -x
 MANTA=$(vmadm lookup alias=manta0)
 
 # Set up the extra manta and manta-nat networks
-ln -s /zones/$MANTA/root/opt/smartdc/manta-deployment/networking /var/tmp/networking
-cd /var/tmp/networking
-./gen-coal.sh > configs/coal.json
-./manta-net.sh ./configs/coal.json
-cd -
+if [[ ! -f /var/tmp/networking/configs/coal.json ]]; then
+    ln -s /zones/$MANTA/root/opt/smartdc/manta-deployment/networking /var/tmp/networking
+    cd /var/tmp/networking
+    ./gen-coal.sh > configs/coal.json
+    ./manta-net.sh ./configs/coal.json
+    cd -
+fi
 
 # Now deploy...
 zlogin $MANTA "bash -l -c \"/opt/smartdc/manta-deployment/bin/manta-init -e nobody@joyent.com -s coal -r coal -m fd2cc906-8938-11e3-beab-4359c665ac99\""
